2003-09-10 23:27 UTC+0200 Przemyslaw Czerpak <druzus@polbox.com>   * xharbour/source/rdd/dbf1.c     * small fix.
  * xharbour/source/rdd/rddsys.prg
    * addede REQUEST DBFDBT
      if user don't want to link with your apps default DBFNTX and DBFFPT
      RDDs then has to add to your sources (like in Clipper)
          ANNOUNCE RDDSYS

  * xharbour/utils/hbtest/hbtest.prg
    * removed unnecessary now REQUEST DBFDBT

  * xharbour/make_tgz.sh
  * xharbour/xharbour.spec
    * don't build fmmt libs if HB_MT != MT

+++ xharbour/make_tgz.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 #
-# $Id: make_tgz.sh,v 1.12 2003/08/20 09:14:41 druzus Exp $
+# $Id: make_tgz.sh,v 1.13 2003/09/08 19:18:42 druzus Exp $
 #
 
 # ---------------------------------------------------------------
@@ -64,9 +64,11 @@ pushd source/vm
     make fm.o
     ar -r $HB_LIB_INSTALL/libfm.a fm.o
     rm -f fm.o
-    make fm.o 'HB_LIBCOMP_MT=YES'
-    ar -r $HB_LIB_INSTALL/libfmmt.a fm.o
-    rm -f fm.o
+    if [ $HB_MT = "MT" ]; then
+        make fm.o 'HB_LIBCOMP_MT=YES'
+        ar -r $HB_LIB_INSTALL/libfmmt.a fm.o
+        rm -f fm.o
+    fi
     C_USR=$TMP_C_USR
 popd
