+++ xharbour/xharbour.spec
@@ -1,5 +1,5 @@
 #
-# $Id: xharbour.spec,v 1.27 2003/08/20 09:14:41 druzus Exp $
+# $Id: xharbour.spec,v 1.28 2003/09/08 19:18:42 druzus Exp $
 #
 
 # ---------------------------------------------------------------
@@ -237,9 +237,11 @@ pushd source/vm
     make fm.o
     ar -r $HB_LIB_INSTALL/libfm.a fm.o
     rm -f fm.o
-    make fm.o 'HB_LIBCOMP_MT=YES'
-    ar -r $HB_LIB_INSTALL/libfmmt.a fm.o
-    rm -f fm.o
+    if [ $HB_MT = "MT" ]; then
+        make fm.o 'HB_LIBCOMP_MT=YES'
+        ar -r $HB_LIB_INSTALL/libfmmt.a fm.o
+        rm -f fm.o
+    fi
     C_USR=$TMP_C_USR
 popd
